filtering and additional calculation
added various lpf, mavg, and exp filters to incoming torque and angle data, as well as the new derivative force output. Derivative force ultimately is going to be what we need to monitor in the final program. Just a matter of cleaning out the forces experienced due to motion/gravity. I don't think smoothing this much is doing us any good.